Several fixes, compatibility with proper jabber presence updates. Now we need to expose status changes globaly in the app. The websocket server will now send a message with an attribute called 'show' (compatible with jabber) which can be:
- xa (voluntary away)
- away (idle)
- chat (available)
- dnd (busy)

And I added 3 more:

- offline
- join-room
- exit-room

For room participants, the presence handler only needs to see if the message show is either join-room or exit-room, the rest of the message types should be for global status updates in the window.

Also, the message contains a status attribute which is not blank when show is one of the first 4 described values.